PR review updates, delete unnecessary comments
