--- projects/lf-documentation/src/app/lf-repository-browser-documentation/lf-repository-browser-documentation.component.spec.ts
@@ -1,25 +1,0 @@
-// import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-// import { LfRepositoryBrowserDocumentationComponent } from './lf-repository-browser-documentation.component';
-
-// describe('LfRepositoryBrowserDocumentationComponent', () => {
-//   let component: LfRepositoryBrowserDocumentationComponent;
-//   let fixture: ComponentFixture<LfRepositoryBrowserDocumentationComponent>;
-
-//   beforeEach(async () => {
-//     await TestBed.configureTestingModule({
-//       declarations: [ LfRepositoryBrowserDocumentationComponent ]
-//     })
-//     .compileComponents();
-//   });
-
-//   beforeEach(() => {
-//     fixture = TestBed.createComponent(LfRepositoryBrowserDocumentationComponent);
-//     component = fixture.componentInstance;
-//     fixture.detectChanges();
-//   });
-
-//   it('should create', () => {
-//     expect(component).toBeTruthy();
-//   });
-// });
