+++ projects/styles/lf-laserfiche-lite/sass/_material.scss
@@ -46,7 +46,3 @@
   -webkit-font-feature-settings: 'liga';
   -webkit-font-smoothing: antialiased;
 }
-
-// .lf-menu-item[mat-menu-item]:hover:not([disabled]) {
-//   background-color: palette-lf.$lf-menu-item-hover-color;
-// }
