--- projects/ui-components/lf-metadata/field-components/field-base-parts/lf-field-base/base-field/base-field.directive.spec.ts
@@ -1,14 +1,0 @@
-// import { waitForAsync, TestBed } from '@angular/core/testing';
-// import { LfFieldTokenService } from '../lf-field-token.service';
-
-// import { BaseFieldDirective } from './base-field.directive';
-
-// describe('BaseField', () => {
-//   beforeEach(waitForAsync(() => {
-//     TestBed.configureTestingModule({
-//       declarations: [BaseFieldDirective],
-//       providers: [LfFieldTokenService]
-//     })
-//       .compileComponents();
-//   }));
-// });
