+++ projects/ui-components/lf-repository-browser/lf-repository-browser.module.ts
@@ -1,39 +1,21 @@
 import { CUSTOM_ELEMENTS_SCHEMA, Injector, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {ScrollingModule} from '@angular/cdk/scrolling';
+import { ScrollingModule } from '@angular/cdk/scrolling';
 import { createCustomElement } from '@angular/elements';
 import { FormsModule } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatButtonToggleModule } from '@angular/material/button-toggle';
-import { MatDialogModule } from '@angular/material/dialog';
-import { MatIconModule } from '@angular/material/icon';
-import { MatListModule } from '@angular/material/list';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatTooltipModule } from '@angular/material/tooltip';
-
 import { LfRepositoryBrowserComponent } from './lf-repository-browser.component';
 import { LfBreadcrumbsModule, LfLoaderModule } from '@laserfiche/lf-ui-components/shared';
 import { LfSelectionListModule } from '@laserfiche/lf-ui-components/lf-selection-list';
 
-
 @NgModule({
-  declarations: [
-    LfRepositoryBrowserComponent
-  ],
+  declarations: [LfRepositoryBrowserComponent],
   imports: [
     CommonModule,
-    MatIconModule,
     FormsModule,
-    MatButtonModule,
-    MatListModule,
-    MatButtonToggleModule,
-    MatMenuModule,
-    MatTooltipModule,
-    MatDialogModule,
     ScrollingModule,
     LfBreadcrumbsModule,
     LfLoaderModule,
-    LfSelectionListModule
+    LfSelectionListModule,
   ],
   bootstrap: [LfRepositoryBrowserComponent],
   exports: [LfRepositoryBrowserComponent],
@@ -41,9 +23,7 @@ import { LfSelectionListModule } from '@laserfiche/lf-ui-components/lf-selection
 })
 export class LfRepositoryBrowserModule {
   /** @internal */
-  constructor(
-    /** @internal */ injector: Injector
-  ) {
+  constructor(/** @internal */ injector: Injector) {
     const repositoryExplorerElementName: string = 'lf-repository-browser';
     if (window.customElements && !customElements.get(repositoryExplorerElementName)) {
       const repoExplorerElement = createCustomElement(LfRepositoryBrowserComponent, { injector });
